Refactor preview process management: add concurrency limits, enhance error handling, implement idle reaper, optimize port allocation, and improve logging with bounded buffers. Transition billing to decimal-based credit handling for precision and introduce new token-based usage charges with configurable markup. Add optional pagination for project listing.

+++ apps/api/app/models/billing.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from app.db.base import Base
-from sqlalchemy import String, DateTime, Text, Integer
+from sqlalchemy import String, DateTime, Text, Integer, Numeric
 from sqlalchemy.orm import Mapped, mapped_column
 
 
@@ -19,7 +19,7 @@ class UserAccount(Base):
     plan: Mapped[str | None] = mapped_column(String(64), nullable=True)  # free, pro, team, enterprise
 
     # Credits
-    credit_balance: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
+    credit_balance: Mapped[float] = mapped_column(Numeric(12, 6), default=0, nullable=False)
 
     # Audit
     updated_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow,
@@ -32,7 +32,7 @@ class CreditTransaction(Base):
     id: Mapped[str] = mapped_column(String(64), primary_key=True)
     owner_id: Mapped[str] = mapped_column(String(128), index=True)
 
-    amount: Mapped[int] = mapped_column(Integer, nullable=False)  # positive for add, negative for spend/refund
+    amount: Mapped[float] = mapped_column(Numeric(12, 6), nullable=False)  # positive for add, negative for spend/refund
     tx_type: Mapped[str] = mapped_column(String(16), nullable=False)  # purchase, spend, refund, grant
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
