+++ apps/api/app/services/users_service.py
@@ -22,7 +22,7 @@ class UserProfileDTO:
     last_login_at: Optional[datetime] = None
     last_active_at: Optional[datetime] = None
     plan: Optional[str] = None
-    credit_balance: Optional[int] = None
+    credit_balance: Optional[float] = None
 
 
 class UsersService:
